+++ init.lua
@@ -18,11 +18,5 @@ vim.opt.rtp:prepend(lazypath)
 vim.loader.enable()
 
 require 'settings'.setup()
-require 'custom_filetypes'.setup()
-require 'lazy-plugins'.setup()
-require 'highlight-parse'.setup()
-require 'completion'.setup()
-require 'lsp-config'.setup()
-require 'lsp-null-ls'.setup()
-require 'mappings'.setup()
-require 'themes'.setup()
+require 'plugins'.setup()
+require 'config'.setup()
